StudentDTO.java:add class dto

+++ src/main/java/com/project/app/dto/StudentDTO.java
@@ -2,7 +2,7 @@ package com.project.app.dto;
 
 public class StudentDTO extends HumanDTO {
 
-    public StudentDTO(int id, int age, String sex, double salary, String type) {
+    public StudentDTO(Long id, int age, String sex, double salary, String type) {
         super(id, age, sex, salary, type);
     }
 }
